+++ packages/backend/src/index.ts
@@ -100,11 +100,9 @@ export type {
   PaginatedResponseJSON,
   TestingTokenJSON,
   WebhooksSvixJSON,
-  CommercePayerJSON,
   CommercePlanJSON,
-  CommerceSubscriptionItemJSON,
-  CommercePaymentAttemptJSON,
   CommerceSubscriptionJSON,
+  CommerceSubscriptionItemJSON,
 } from './api/resources/JSON';
 
 /**
@@ -147,6 +145,8 @@ export type {
   User,
   TestingToken,
   CommercePlan,
+  CommerceSubscription,
+  CommerceSubscriptionItem,
 } from './api/resources';
 
 /**
@@ -166,6 +166,9 @@ export type {
   WaitlistEntryWebhookEvent,
   WebhookEvent,
   WebhookEventType,
+  CommercePaymentAttemptWebhookEvent,
+  CommerceSubscriptionWebhookEvent,
+  CommerceSubscriptionItemWebhookEvent,
 } from './api/resources/Webhooks';
 
 /**